Entities + Models

+++ src/main/java/com/spring/henallux/firstSpringProject/Model/CartLine.java
@@ -29,7 +29,7 @@ public class CartLine {
         this.quantity = quantity;
     }
 
-    public float getTotalPrice() {
+    public Double getTotalPrice() {
         return product.getPrice() * quantity;
     }
